+++ src/Helpers/SpellHelper.php
@@ -173,7 +173,8 @@ class SpellHelper
             'self_spell_duration' => 'but increases duration by %d%%',
             'convert_military_spies_to_military_draftees' => 'Turns %g%% of enemy spies into draftees',
             'convert_peasants_to_self_military_unit3' => 'Kills %g%% of enemy peasants, converting 5%% into Progeny',
-            'destroy_peasants' => 'Kills %g%% peasants',
+            'apply_burning' => '%g%% chance to inflict Burning if at war',
+            'destroy_peasants' => 'Kills %g%% unprotected peasants',
             'destroy_resource_food' => 'Destroys %g%% crops',
             'destroy_improvement_science' => 'Destroys %g%% science',
             'destroy_improvement_keep' => 'Destroys %g%% keep',
@@ -188,6 +189,11 @@ class SpellHelper
             'scale_by_day' => 'Scales by day in round from 137.5%% to 62.5%%',
             'explore_cost_wizard_mastery' => 'Exploring platinum cost reduced by 1%% per %d Wizard Mastery',
             'spell_refund' => 'Failed spells refund %d%% of their mana cost',
+            'fireball_vulnerability' => 'Increases vulnerability to fireball by %g%% of your maximum peasant population per stack (max 3)',
+            'fixed_population_growth' => 'Population growth is fixed at %g%% of your vulnerable peasant population',
+            'apply_rejuvination' => 'Applies Rejuvenation upon expiration',
+            'enemy_fireball_damage' => '%+g%% enemy fireball damage',
+            'immunity_burning' => 'Immune to Burning',
         ];
     }
 
@@ -223,10 +229,11 @@ class SpellHelper
     public function getCategoryString(string $category) {
         $categories = [
             'info' => 'Information',
-            'self' => 'Self',
             'hostile' => 'Offensive',
             'war' => 'War',
             'wonder' => 'Wonder',
+            'self' => 'Self',
+            'effect' => 'Status Effect',
         ];
 
         return $categories[$category];
@@ -234,6 +241,27 @@ class SpellHelper
 
     public function getSpellType(Spell $spell) {
         return $this->getCategoryString($spell->category);
+    }
+
+    public function getStatusEffectStacks(array $applications) {
+        if (isset($applications['realm_ids'])) {
+            return count($applications['realm_ids']);
+        }
+        return 0;
+    }
+
+    public function getStatusEffectStacksDisplay(array $activeSpell) {
+        if (isset($activeSpell['applications'])) {
+            $applications = $activeSpell['applications'];
+            if (!is_array($applications)) {
+                $applications = json_decode($applications, JSON_OBJECT_AS_ARRAY);
+            }
+            $stacks = $this->getStatusEffectStacks($applications);
+            if ($stacks > 1) {
+                return " - {$stacks}x";
+            }
+        }
+        return '';
     }
 
     public function obfuscateInfoOps(array $infoOps) {