+++ CHANGELOG.md
@@ -15,7 +15,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
 ### Changed
 - A percentage of your peasant population is now invulnerable to fireball damage (scales down during the round from 90% on Day 4 to 70% on Day 44)
 - Up to 50% of your vulnerable peasant population is protected by defensive WPA
-- Fireball: now kills 5% of your target's unprotected peasant population (from 2.75% of current peasants) and 2% of your target's food (from 2.75%)
+- Fireball: now kills 5% of your target's maximum unprotected peasant population (from 2.75% of current peasants) and 2% of your target's food (from 2.75%), 30% chance to apply Burning if at war
+- Burning: new status effect (spell) with 36 hour duration, increases vulnerability to fireball by 10% of your maximum peasant population (20% in mutual war), population growth is fixed at 5% of your vulerable peasant population, stacks if applied by another realm to a maximum of 30%, applies Rejuvenation upon expiration
 
 ## [1.37.2] - 2023-10-24
 ### Added